organize layers in pack command
