+++ package.json
@@ -33,7 +33,6 @@
     "format:check": "oxfmt --check",
     "lint-lockfile": "lint-lockfile",
     "lint-repo": "node ./scripts/src/tasks/lintRepo.ts",
-    "organize-packs": "node ./scripts/src/cli.ts organize-packs",
     "test-links": "markdown-link-check"
   },
   "devDependencies": {
