+++ scripts/src/cli.ts
@@ -6,7 +6,6 @@ import { FormatCommand } from './tasks/format.ts';
 import { LintCommand } from './tasks/lint.ts';
 import { JestCommand } from './tasks/jest.ts';
 import { CheckPublishingCommand } from './tasks/checkPublishingTask.ts';
-import { OrganizePacksCommand } from './tasks/organizePacks.ts';
 
 const cli = new Cli({
   binaryLabel: 'fluentui-scripts CLI',
@@ -19,7 +18,6 @@ cli.register(FormatCommand);
 cli.register(LintCommand);
 cli.register(JestCommand);
 cli.register(CheckPublishingCommand);
-cli.register(OrganizePacksCommand);
 
 cli.register(Builtins.DefinitionsCommand);
 cli.register(Builtins.HelpCommand);
